# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for app.",
        "title": "perna",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/domestic-violence-bos": {
            "get": {
                "description": "Get all Domestic Violence BOs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomesticViolenceBO"
                ],
                "summary": "Get All Domestic Violence BOs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.DomesticViolenceBO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Domestic Violence BO record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomesticViolenceBO"
                ],
                "summary": "Create Domestic Violence BO",
                "parameters": [
                    {
                        "description": "Domestic Violence BO Data",
                        "name": "DomesticViolenceBO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.DomesticViolenceBO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.DomesticViolenceBO"
                        }
                    }
                }
            }
        },
        "/api/domestic-violence-bos/{id}": {
            "get": {
                "description": "Get a single Domestic Violence BO by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomesticViolenceBO"
                ],
                "summary": "Get Domestic Violence BO by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Domestic Violence BO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.DomesticViolenceBO"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a Domestic Violence BO record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomesticViolenceBO"
                ],
                "summary": "Update Domestic Violence BO by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Domestic Violence BO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Domestic Violence BO Data",
                        "name": "DomesticViolenceBO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.DomesticViolenceBO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.DomesticViolenceBO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Domestic Violence BO record by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomesticViolenceBO"
                ],
                "summary": "Delete Domestic Violence BO by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Domestic Violence BO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted Successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.DomesticViolenceBO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date_of_incident": {
                    "description": "Data do incidente",
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incident_address": {
                    "description": "Local do incidente",
                    "type": "string"
                },
                "offender_name": {
                    "description": "Nome do agressor",
                    "type": "string"
                },
                "police_station": {
                    "description": "Delegacia responsável",
                    "type": "string"
                },
                "report_details": {
                    "description": "Detalhes do boletim de ocorrência",
                    "type": "string"
                },
                "status": {
                    "description": "Status do boletim (ex: Em andamento, Concluído)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "victim_contact": {
                    "description": "Contato da vítima",
                    "type": "string"
                },
                "victim_name": {
                    "description": "Nome da vítima",
                    "type": "string"
                }
            }
        },
...
        "interfaces.DomesticViolenceBO": {
            "description": "Registro de B.O de Violência Doméstica contra Mulher.",
            "type": "object",
            "properties": {
                "date_of_incident": {
                    "description": "Data do incidente",
                    "type": "string"
                },
                "incident_address": {
                    "description": "Local do incidente",
                    "type": "string"
                },
                "offender_name": {
                    "description": "Nome do agressor",
                    "type": "string"
                },
                "police_station": {
                    "description": "Delegacia responsável",
                    "type": "string"
                },
                "report_details": {
                    "description": "Detalhes do boletim de ocorrência",
                    "type": "string"
                },
                "status": {
                    "description": "Status do boletim (ex: Em andamento, Concluído)",
                    "type": "string"
                },
                "victim_contact": {
                    "description": "Contato da vítima",
                    "type": "string"
                },
                "victim_name": {
                    "description": "Nome da vítima",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}